fix the bottom-line width & spacing of the titles
The distance between section titles and their bottom lines was a
bit longer. Previously I tried to reduce the space by changing the
position of the bottom lines a little bit upper than their original
position. But this is not a conventional way to do this. Finally,
I found that the line height is the reason for this issue. So I set
the line height for the headings to zero. I also set the width of
the bottom line 1px.