fix(team): the image bottom spacing issue
The image overlay in the hover state was showing a little bit longer
than the image. Previously I tried to fix this by changing the overlay
position to 3px at the bottom though it is not a good practice. I
found the reason now. The line-height issue again. I changed the
line-height and position bottom to zero.